test(eventhubs): cover the receive span with a live test
The receive span of PartitionClient::ReceiveEvents has no offline
coverage. A partition client comes only from a factory that opens a
live AMQP link, so no unit test can reach the span, the message count,
or the exception path.

This test runs against a real namespace. It gives the consumer client
a recording tracing provider, receives events from a partition, and
makes sure the span name, the span kind, the messaging attributes, and
the message count are correct. The count assertion compares the span
attribute with the size of the vector that ReceiveEvents returned.

The recording tracing test doubles move from eventhubs_tracing_test.cpp
into a shared header, because the live test lives in a different file.
